Added a cluster page, the overview currently works
The implementation should pick up different clusters accross different
hosts that we manage, for example;

If we manage 6 hosts and;

3 are in "cluster 0"
3 are in "cluster 1"

Then that should be recognised and split up into multiple clusters.

Currently the clusters aren't stored in the database so we are stuck
with incrementing by 1 cluster names.

The way we currently group hosts into clusters is by iterating all our
hosts and asking if its in a cluster and what members it has, if a host
has been seen previously in a cluster then it isn't asked about members
and is skipped (although added to the prev cluster group when its found
), this means that a host can only be part of 1 cluster so if LXD
changes their code to allow an host to be part of multiple clusters
this will need updating