Fixed sidebar lost when LXD host goes offline (close #152)
We check every minute for hosts that are still online, if you happen
to be using LXDMosaic in a period where a host is down but the cron job
hasn't run - it locks and you cant do anything.

This commit introduces an attempt to parse exception messages to
detect a particular curl error message, extract host details and
disable the host (if we can find it by url/ip), its messy and discusting
but so is leaving the user unable todo anything for upto 1 minute!

It also introduces the need for annotations and late property injection
into the execption handaler, which I really dont like, but there isn't
much other choice, without alot more work!