Fix backing up container in snap (close #264)
When you use SNAP you are forced to use SNAP_COMMON to store your backups for
your instances - if we use any other directory it will be versioned (which
will use tons of space)

PHP in the snap wasn't importing the "external" environment variables (like
SNAP_COMMON) so ive modified the snap php.ini to ensure it imports the
environment.

There seems to be a weird bug with snap with the cron job, it reports the
following on first few runs;

"snap.lxdmosaic.lxdmosaic-cron.service: Failed to reset devices.list:
Operation not permitted"

Then it resolves itself(?) and starts working (backups schedules run etc),

I dont think this is "blocking" as it could take a while to get to the bottom of
+ we have cron jobs that run frequently which will clear the above errors,
I.E A user would have to setup a backup job in the first 10 minutes after
installation & then schedule that backup < 10 minutes away from installation to
really notice the effects of this!