Make backup path hostId/project/instance instead of hostId/instance
This might lead to some weird looking temporary backup structures
but they  should resolve themselves if they are managed by backup
schedules or as the user deletes the backups.

For example theres a "weird structure" if you have backed up a instance
called "test" & have a project called "test". If you now start backing
up instances in the project test you dir will structure will look
something like this;

     hostId/
       test/
          instanceName/
          backup_file.gz

If the backup is scheduled then eventually the backup_file backup
file will be removed, all the new backups will be under instanceName/ so
this situation resolves itself.

If the backups are manually made then it will remain until deleted by
the user, not really a problem but if the dir tree was to manually
investigated by the user it might look weird.