Fix removing Host from HostCollection causing timeouts
PHP doesn't appear to like when you use $this to start iterating
an object from within iteslf and then deleting an element.
So instead we directly loop over the private array from within the class
and remove the element that way, might be a bug in PHP might be bad code

A quick google search doesn't yeild many anwsers so just going to move
on as this appears to fix it!