Disable the spice channels usbredir, webdav, & cursor to improve UX
webdav & usbredir aren't implemented in spicehtml5 so instead of trying
to open a socket and waiting for a time out we skip the all-together.
Opening the socket seemed to fail more often that not with a 1006 error
or timeout which took far to long, blocking the other sockets from
initiating in a timely manner.

cursor is supported by spicehtml5 but the socket fails with a 1006 and
blocks the inputs socket while it loads. From the spice docs its purpose
is "controlling cursor shape and position on the remote display area"
which as we currently only allow 1 user at a time, this isn't critical.

Removing cursor is a cheap work around but the 1006 socket errors are
currently proving hard to debug in either chrome or firefox so this will
have todo.

I suspect something is wrong with our node proxy because you will
occasionally get 1006 websocket errors on the inputs channel which
prevents the user from interacting with the VM, interestingly you
only get the 1006 error on the inputs channel it doesn't happen with
"main" or "display". I spent a few hours trying to debug it but couldn't
find the source.

Overall it works better but still needs more work.

close #463