Use fbirlik/sqlite3 PPA for later version of sqlite3 in snap build
This is required because PHP 7.4 stops shipping the v3.28.0 libsqlite3
in its build files and relies on the system instead.

The system version in ubuntu core18 is v3.22.0 but we require atleast
version v3.24.0 for UPSERT support.

This is a hacky fix, it would be better to build ourselves and control
the version but master / prod is broken so it will have todo.

The PPA is currently shipping v3.35.5 so we should be careful to
build that same minor version when building ourselves.

close #507