Remove deprecated /hosts/message/ route from node
This was a real security vunreabilty, the risk was mitigated in a
previous release this just removes the code still calling it (the route
responded with success even though it does nothing now).

If this route was to be re-introduced it should only be available over
a socket accessable to PHP, not the network as it was.