Remove deprecated /deploymentProgress/:deploymentId route from event.js
Removes;
  - `vendor` phone-home string from deployments
  - deploymentProgress/:deploymentId route from event.js

Its a route that doesn't require authentication that could be abused,
really it should be allowed to stay but this commit removes all the code
at once so it could be undone.

The rational behind it is that once a cloud-init is done running
its an endpoint that can be called to say "yes im done setting up".

Its a route that cant really ever have auth on it, but;
 - We have no security docs warning about it (its not a real vuln)
 - Deployments is kinda dead, it needs an overhall
 - The code it called had been removed before and this was just a stub

This is a shame because I remeber the pain of getting it working, but
that will teach me to push code that isn't ready!